added checkLives function to separate file
